refactor(pages): rename pageheaders for helmet
