+++ src/App.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react'
 import Routes from './Routes'
-import { Head } from './components'
+import { PageHeaders } from './components'
 
 export default class App extends Component {
   render() {
     return (
       <div className="Application">
-        <Head />
+        <PageHeaders />
         <Routes />
       </div>
     )
