--- src/components/Head/index.js
@@ -1,2 +1,0 @@
-import Head from './Head'
-export { Head }
