+++ src/components/PageHeaders/index.js
@@ -1,0 +1,2 @@
+import PageHeaders from './PageHeaders'
+export { PageHeaders }
