+++ src/components/index.js
@@ -1,4 +1,4 @@
-import { Head } from './Head'
+import { PageHeaders } from './PageHeaders'
 import { MainNav, SocialNav } from './Nav'
 import { Footer } from './Footer'
 import { Background } from './Background'
@@ -6,7 +6,7 @@ import { Background } from './Background'
 import { ImageLoader, StylesHelper } from './Utils/'
 
 export {
-  Head,
+  PageHeaders,
   MainNav,
   SocialNav,
   Footer,
